intersection logic is put in a separate helper file
