+++ src/components/SvgVennDiagram4.tsx
@@ -6,11 +6,12 @@ import { useQuery } from '../helpers/query-hook';
 import { useExploreUrl } from '../helpers/explore-url';
 import { MutationProportionData } from '../data/MutationProportionDataset';
 import Loader from './Loader';
-import _ from 'lodash';
+// import _ from 'lodash';
 import './style/svgPlots.css';
 import { ReferenceGenomeService } from '../services/ReferenceGenomeService';
 import { OverlayTrigger, Tooltip } from 'react-bootstrap';
 import Slider from 'rc-slider';
+import { intersections4 } from '../helpers/intersections';
 
 import OutlinedInput from '@mui/material/OutlinedInput';
 import InputLabel from '@mui/material/InputLabel';
@@ -89,355 +90,7 @@ export const SvgVennDiagram4 = ({ selectors }: Props) => {
         .filter(m => m.proportion >= minProportion)
         .map(m => m.mutation);
 
-      const shared = _.intersection(
-        variant1Mutations,
-        variant2Mutations,
-        variant3Mutations,
-        variant4Mutations
-      );
-
-      let shared1and2 = _.intersection(variant1Mutations, variant2Mutations); // const shared1and2 = _.pullAll(shared1and2, shared)
-      let shared1and3 = _.intersection(variant1Mutations, variant3Mutations);
-      let shared1and4 = _.intersection(variant1Mutations, variant4Mutations);
-      let shared2and3 = _.intersection(variant2Mutations, variant3Mutations);
-      let shared2and4 = _.intersection(variant2Mutations, variant4Mutations);
-      let shared3and4 = _.intersection(variant3Mutations, variant4Mutations);
-
-      let shared1and2and3 = _.intersection(variant1Mutations, variant2Mutations, variant3Mutations);
-
-      let shared1and2and4 = _.intersection(variant1Mutations, variant2Mutations, variant4Mutations);
-      let shared2and3and4 = _.intersection(variant4Mutations, variant2Mutations, variant3Mutations);
-      let shared1and4and3 = _.intersection(variant4Mutations, variant1Mutations, variant3Mutations);
-
-      shared1and2 = _.pullAll(shared1and2, [
-        ...new Set(
-          [
-            shared,
-            shared1and3,
-            shared2and3,
-            shared1and4,
-            shared2and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-      shared1and3 = _.pullAll(shared1and3, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared2and3,
-            shared1and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      shared1and4 = _.pullAll(shared1and4, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared2and3,
-            shared2and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      shared2and3 = _.pullAll(shared2and3, [
-        ...new Set(
-          [
-            shared,
-            shared1and3,
-            shared1and2,
-            shared2and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      shared2and4 = _.pullAll(shared2and4, [
-        ...new Set(
-          [
-            shared,
-            shared2and3,
-            shared1and2,
-            shared3and4,
-            shared1and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      shared3and4 = _.pullAll(shared2and3, [
-        ...new Set(
-          [
-            shared,
-            shared1and3,
-            shared1and2,
-            shared2and4,
-            shared2and3,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      shared1and2and3 = _.pullAll(shared1and2and3, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared1and3,
-            shared1and4,
-            shared2and3,
-            shared2and4,
-            shared3and4,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-      shared1and2and4 = _.pullAll(shared1and2and4, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared1and3,
-            shared1and4,
-            shared2and3,
-            shared2and4,
-            shared3and4,
-            shared1and2and3,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-      shared2and3and4 = _.pullAll(shared2and3and4, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared1and3,
-            shared1and4,
-            shared2and3,
-            shared2and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-      shared1and4and3 = _.pullAll(shared1and4and3, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared1and3,
-            shared1and4,
-            shared2and3,
-            shared2and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-          ].flat()
-        ),
-      ]);
-
-      const onlyVariant1 = _.pullAll(variant1Mutations, [
-        ...new Set(
-          [
-            shared,
-            shared1and2,
-            shared1and3,
-            shared1and4,
-            shared2and3,
-            shared2and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-      const onlyVariant2 = _.pullAll(variant2Mutations, [
-        ...new Set(
-          [
-            shared,
-            shared2and3,
-            shared2and4,
-            shared1and2,
-            shared1and3,
-            shared1and4,
-            shared3and4,
-            shared1and2and3,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      const onlyVariant3 = _.pullAll(variant3Mutations, [
-        ...new Set(
-          [
-            shared,
-            shared1and3,
-            shared2and3,
-            shared3and4,
-            shared1and4,
-            shared1and2,
-            shared2and4,
-            shared1and2and3,
-            shared2and3and4,
-            shared1and4and3,
-            shared1and2and4,
-          ].flat()
-        ),
-      ]);
-      const onlyVariant4 = _.pullAll(variant4Mutations, [
-        ...new Set(
-          [
-            shared,
-            shared1and4,
-            shared2and4,
-            shared3and4,
-            shared1and2and4,
-            shared2and3and4,
-            shared1and4and3,
-          ].flat()
-        ),
-      ]);
-
-      // Group by genes
-      const genes = new Map<
-        string,
-        {
-          onlyVariant1: string[];
-          onlyVariant2: string[];
-          onlyVariant3: string[];
-          onlyVariant4: string[];
-          shared: string[];
-          shared1and2: string[];
-          shared1and3: string[];
-          shared1and4: string[];
-          shared2and3: string[];
-          shared2and4: string[];
-          shared3and4: string[];
-          shared1and2and3: string[];
-          shared1and2and4: string[];
-          shared2and3and4: string[];
-          shared1and4and3: string[];
-        }
-      >();
-
-      for (let gene of ReferenceGenomeService.genes) {
-        genes.set(gene, {
-          onlyVariant1: [],
-          onlyVariant2: [],
-          onlyVariant3: [],
-          onlyVariant4: [],
-          shared: [],
-          shared1and2: [],
-          shared1and3: [],
-          shared1and4: [],
-          shared2and3: [],
-          shared2and4: [],
-          shared3and4: [],
-          shared1and2and3: [],
-          shared1and2and4: [],
-          shared2and3and4: [],
-          shared1and4and3: [],
-        });
-      }
-
-      for (let mutation of shared) {
-        genes.get(mutation.split(':')[0])!.shared.push(mutation);
-      }
-      for (let mutation of shared1and2) {
-        genes.get(mutation.split(':')[0])!.shared1and2.push(mutation);
-      }
-      for (let mutation of shared1and3) {
-        genes.get(mutation.split(':')[0])!.shared1and3.push(mutation);
-      }
-
-      for (let mutation of shared1and4) {
-        genes.get(mutation.split(':')[0])!.shared1and4.push(mutation);
-      }
-
-      for (let mutation of shared2and3) {
-        genes.get(mutation.split(':')[0])!.shared2and3.push(mutation);
-      }
-
-      for (let mutation of shared2and4) {
-        genes.get(mutation.split(':')[0])!.shared2and4.push(mutation);
-      }
-
-      for (let mutation of shared3and4) {
-        genes.get(mutation.split(':')[0])!.shared3and4.push(mutation);
-      }
-
-      for (let mutation of shared1and2and3) {
-        genes.get(mutation.split(':')[0])!.shared1and2and3.push(mutation);
-      }
-
-      for (let mutation of shared1and2and4) {
-        genes.get(mutation.split(':')[0])!.shared1and2and4.push(mutation);
-      }
-
-      for (let mutation of shared1and4and3) {
-        genes.get(mutation.split(':')[0])!.shared1and4and3.push(mutation);
-      }
-
-      for (let mutation of shared2and3and4) {
-        genes.get(mutation.split(':')[0])!.shared2and3and4.push(mutation);
-      }
-
-      for (let mutation of onlyVariant1) {
-        genes.get(mutation.split(':')[0])!.onlyVariant1.push(mutation);
-      }
-      for (let mutation of onlyVariant2) {
-        genes.get(mutation.split(':')[0])!.onlyVariant2.push(mutation);
-      }
-      for (let mutation of onlyVariant3) {
-        genes.get(mutation.split(':')[0])!.onlyVariant3.push(mutation);
-      }
-
-      for (let mutation of onlyVariant4) {
-        genes.get(mutation.split(':')[0])!.onlyVariant4.push(mutation);
-      }
-
-      // Count
-      return [...genes.entries()].map(([gene, muts]) => ({
-        gene,
-        ...muts,
-      }));
+      return intersections4(variant1Mutations, variant2Mutations, variant3Mutations, variant4Mutations);
     }
   }, [res.data, minProportion]);
